Load programs and data file
Add the java source and compiled program for TicTacToe.  It also comes with a text file loserecords.txt.  

If we run the program with the text file in the same directory, the program will use the "loserecords.txt" to load the computer's previous game loses.  With this, it will decrease the chance of the computer's losing the game.  In other words, if we play the game with the text file, the computer already has some lose experience, and therefore, it is smarter not making the same lose again.

If we play the game without the text file, the computer will start as fresh, and learn from the lose gradually.